cachefiles: allocate per-subdir content map files
Allocate one content map file for each sub-directory under one volume,
so that the cacehfilesd only needs to remove the whole sub-directory
(including the content map file and backing files in the sub-directory)
as usual when it's going to cull the whole sub-directory or volume.

The content map file will be shared among all backing files under this
same sub-directory.

Signed-off-by: Jingbo Xu <jefflexu@linux.alibaba.com>