erofs: remove unused device mapping in the meta routine
Currently there're two anonymous inodes (inode and anon_inode in struct
erofs_fscache) may be used for each blob.  The former is only for
bootstrap and used as the address_space of page cache, while the latter
is used for both bootstrap and data blobs when share domain mode enabled
and behaves as a sentinel in the shared domain.

In prep for the following support for page cache sharing, following
patch will unify these two anonymous inodes.  That is, the unified
anonymous inode not only acts as the address_space of page cache, but
also a sentinel in share domain mode.

However the current meta routine can't work if above change applied.
Current meta routine will make a device mapping, and superblock of the
filesystem is required to do the device mapping.  Currently the
superblock is derived from the input meta folio, which is reasonable
since the anonymous inode (used for the address_space of page cache) is
always allocated from the filesystem's sb.  However after anonymous
inodes are unified, that is no longer always true.  For example, in
share domain mode, the unified anonymous inode will be allocated from
pseudo mnt, and the superblock derived from the folio is actually a
pseudo sb, which can't be used for the device mapping at all.

As for the meta routine itself, currently metadata is always on
bootstrap, which means device mapping is not needed so far.  After
removing the redundant device mapping logic, we can derive the required
fscache_ctx from anonymous inode's i_private.

Signed-off-by: Jingbo Xu <jefflexu@linux.alibaba.com>