composefs: Add filesystem implementation
This is the basic inode and filesystem implementation.

We open a private mount for the specified base directories at the time
of mount, and all backing data files are looked up with this mount as
root, to protect against changes in the mount tables, or links pointing
outside the base directories.

Access to the backing data is done with the callers credentials, to
the backing files need to be readable by all users of the
filesystem. We also use open_with_fake_path() to ensure the backing
filename is not visible in /proc/.

Signed-off-by: Alexander Larsson <alexl@redhat.com>
Co-developed-by: Giuseppe Scrivano <gscrivan@redhat.com>
Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>